+++ CHANGELOG.md
@@ -78,6 +78,9 @@
   This allows to use `Batch` instances with non-standard datatypes.
 * Marked support for non-`Batch` inputs to `augment_batches()` as deprecated.
 * Added `Augmenter.augment()` method.
+* Added `dtypes.clip_()` function.
+* Fixed an issue in `dtypes.clip_to_value_range_()` causing errors when clip value range exceeded array dtype's value range.
+* Fixed an issue in `dtypes.clip_to_value_range_()` when the input array was scalar, i.e. had shape `()`.
 
 
 # 0.2.8